Module lasm::asm

asm, the module that provides structures that represent assembly instructions

Enums

Instruct
This enum represents a single instruction that can be assembled. The majority of the work to implement Target for a target programming language is defining how to convert an instruction to a properly formatted string for the target.

Literal
The Literal enum represents a literal in an assembly program (duh). A literal can either be a double precision float, or a character literal

Register
The Register enum represents a register in an assembly program (obviously)

Constants

DEFAULT_STACK_SIZE
This is the stack size that is used if the assembly file does not specify one.

PREDEFINED_REGISTERS
This specifies the number of predefined registers. This is VERY important to get right, if this is too small, user defined registers will overwrite the Accumulator, StackPointer, and other predefined registers.