Module lasm::target

target, the module responsible for providing common targets for lasm

As stated in the top level documentation of the crate, the purpose of lasm is to be as portable as possible. To maximize portability, the target module provides the Target trait and a few builtin implementations for common programming languages.

Using the Instruct enum in the asm module, though, Target can be implemented for other programming languages by other crates that use this library. Additionally, uses can write more optimized implementations for languages that already have one.

Structs

C
C is a target

Traits

Target
This trait should be implemented for a struct that represents a target language that lasm assembles to.